adc variables are type float, vcoeff correct

+++ firmware/src/adc.h
@@ -22,7 +22,7 @@
 typedef enum adc_channels{ADC1, ADC2, ADC3} adc_channels_t;                           //*< the adc_channel type
 extern volatile uint8_t print_adc;
 
-extern volatile uint16_t batvoltage, position, batcurrent;
+extern volatile float batvoltage, position, batcurrent;
 
 void adc_init(void);
